Refactor StaffPacketHandler
Object -> top level methods + change package
